refactor[#1] : 상위 디렉토리를 feature와 global로 세팅
