+++ build.gradle
@@ -17,6 +17,9 @@ repositories {
 
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+	compileOnly 'org.projectlombok:lombok'
+	implementation 'org.springdoc:springdoc-openapi-ui:1.7.0'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
