--- src/main/java/com/example/kakao_clone/auth/application/AuthService.java
@@ -1,4 +1,0 @@
-package com.example.kakao_clone.auth.application;
-
-public class AuthService {
-}
