--- src/main/java/com/example/kakao_clone/auth/presentation/AuthController.java
@@ -1,4 +1,0 @@
-package com.example.kakao_clone.auth.presentation;
-
-public class AuthController {
-}
