--- src/main/java/com/example/kakao_clone/exception/CustomException.java
@@ -1,13 +1,0 @@
-package com.example.kakao_clone.exception;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException{
-    private ErrorCode errorCode;
-
-    public CustomException(ErrorCode errorCode, String message){
-        super(message);
-        this.errorCode = errorCode;
-    }
-}
