+++ src/main/java/com/example/kakao_clone/feature/user/application/UserService.java
@@ -1,0 +1,12 @@
+package com.example.kakao_clone.feature.user.application;
+
+import com.example.kakao_clone.feature.user.domain.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+}
