--- src/main/java/com/example/kakao_clone/friend/domain/Friend.java
@@ -1,24 +1,0 @@
-package com.example.kakao_clone.friend.domain;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Friend {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String kakaoId;
-
-    private String friendKakaoId;
-
-    private String friendName;
-}
