--- src/main/java/com/example/kakao_clone/friend/domain/repository/FriendRepository.java
@@ -1,9 +1,0 @@
-package com.example.kakao_clone.friend.domain.repository;
-
-import com.example.kakao_clone.friend.domain.Friend;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FriendRepository extends JpaRepository<Friend, Long> {
-}
