+++ src/main/java/com/example/kakao_clone/global/auth/application/AuthService.java
@@ -1,0 +1,4 @@
+package com.example.kakao_clone.global.auth.application;
+
+public class AuthService {
+}
