+++ src/main/java/com/example/kakao_clone/global/auth/presentation/AuthController.java
@@ -1,0 +1,4 @@
+package com.example.kakao_clone.global.auth.presentation;
+
+public class AuthController {
+}
