+++ src/main/java/com/example/kakao_clone/global/common/BaseEntity.java
@@ -1,0 +1,33 @@
+package com.example.kakao_clone.global.common;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.Column;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+public class BaseEntity {
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+    @CreatedDate
+    @Column(updatable = false)
+    private LocalDateTime createDate;
+
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+    @LastModifiedDate
+    private LocalDateTime modifiedDate;
+
+    private boolean deleted = Boolean.FALSE;
+
+    public void deleteContent() {
+        deleted = true;
+    }
+}
