+++ src/main/java/com/example/kakao_clone/global/docs/ApiDoc.md
@@ -1,0 +1,1 @@
+## API 명세
