+++ src/main/java/com/example/kakao_clone/global/exception/CustomException.java
@@ -1,0 +1,13 @@
+package com.example.kakao_clone.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException{
+    private ErrorCode errorCode;
+
+    public CustomException(ErrorCode errorCode, String message){
+        super(message);
+        this.errorCode = errorCode;
+    }
+}
