--- src/main/java/com/example/kakao_clone/room/domain/Room.java
@@ -1,23 +1,0 @@
-package com.example.kakao_clone.room.domain;
-
-import com.example.kakao_clone.common.BaseEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Room extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Integer roomType;
-
-    private String lastChat;
-}
