--- src/main/java/com/example/kakao_clone/user/application/UserService.java
@@ -1,12 +1,0 @@
-package com.example.kakao_clone.user.application;
-
-import com.example.kakao_clone.user.domain.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-
-    private final UserRepository userRepository;
-}
