--- src/main/java/com/example/kakao_clone/user/domain/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.kakao_clone.user.domain.repository;
-
-import com.example.kakao_clone.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
