--- src/main/java/com/example/kakao_clone/user/presentation/UserController.java
@@ -1,12 +1,0 @@
-package com.example.kakao_clone.user.presentation;
-
-import com.example.kakao_clone.user.application.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class UserController {
-
-    private final UserService userService;
-}
